# Handle map { ... }.sum in Performance/Sum