--- a/docs/modules/ROOT/pages/cops_performance.adoc
+++ b/docs/modules/ROOT/pages/cops_performance.adoc
@@ -1651,11 +1651,12 @@ in some Enumerable object can be replaced by `Enumerable#sum` method.
 [1, 2, 3].reduce(10, :+)
 [1, 2, 3].inject(&:+)
 [1, 2, 3].reduce { |acc, elem| acc + elem }
+[1, 2, 3].map { |elem| elem ** 2 }.sum
 
 # good
 [1, 2, 3].sum
 [1, 2, 3].sum(10)
-[1, 2, 3].sum
+[1, 2, 3].sum { |elem| elem ** 2 }
 ----
 
 === References