@shepmaster 2020-09-21
This looks to have issues with `foo.map(&:bar).sum(initial_value)`, which is corrected to `foo.sum(initial_value)(&:bar)`