@eugeneius 2020-09-21
Thanks for reporting that @shepmaster, I've opened https://github.com/rubocop-hq/rubocop-performance/pull/172 to fix it.