@marcandre 2020-09-22
> This looks to have a false positive when used with Active Record:
> 
> ```ruby
> User.limit(10).map(&:id).sum(10)
> # => 141467
> 
> User.limit(10).sum(10, &:id)
> # ArgumentError: Column name argument is not supported when a block is passed.
> # from .../activerecord-6.0.3.3/lib/active_record/relation/calculations.rb:87:in `sum'
> ```

More like an unsafe correction. The code should still be modified, e.g. `User.limit(10).sum(:id) + 10`